+++ wrappers/src/sync_session_cs.cpp
@@ -35,7 +35,7 @@ namespace binding {
     void (*s_refresh_access_token_callback)(std::shared_ptr<SyncSession>*);
     void (*s_session_error_callback)(std::shared_ptr<SyncSession>*, int32_t error_code, const char* message, size_t message_len, std::pair<char*, char*>* user_info_pairs, int user_info_pairs_len);
     void (*s_progress_callback)(size_t, uint64_t transferred_bytes, uint64_t transferrable_bytes);
-    void (*s_wait_callback)(void* task_completion_source, int32_t error_code);
+    void (*s_wait_callback)(void* task_completion_source, int32_t error_code, const char* message, size_t message_len);
 
     void bind_session(const std::string&, const realm::SyncConfig& config, std::shared_ptr<SyncSession> session)
     {
@@ -170,7 +170,7 @@ REALM_EXPORT bool realm_syncsession_wait(const SharedSyncSession& session, void*
 {
     return handle_errors(ex, [&] {
         auto waiter = [task_completion_source](std::error_code error) {
-            s_wait_callback(task_completion_source, error.value());
+            s_wait_callback(task_completion_source, error.value(), error.message().c_str(), error.message().length());
         };
         
         if (direction == CSharpNotifierType::Upload) {